requirements.txt
streamlit==1.24.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
xgboost==1.7.5
requests==2.31.0
python-dotenv==1.0.0
plotly==5.15.0
geopy==2.3.0
